requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.1
defusedxml==0.8.0rc2
dj-rest-auth==6.0.0
Django==5.1.1
django-allauth==64.2.0
django-ckeditor==6.7.1
django-cors-headers==4.4.0
django-js-asset==2.2.0
django-solo==2.3.0
django-taggit==6.0.0
djangorestframework==3.15.2
idna==3.8
oauthlib==3.2.2
pillow==10.4.0
pycparser==2.22
PyJWT==2.9.0
python3-openid==3.2.0
razorpay==1.4.2
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==74.1.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
urllib3==2.2.2
django-background-tasks
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
